requirements.txt

-r requirements-zero.txt
-r requirements-account.txt

pinax-announcements==3.0.0
pinax-images==3.0.0
pinax-likes==3.0.0
pinax-messages==2.0.1
pinax-testimonials==2.0.3

django-test-plus>=1.0.14
